Disable .erb linter and add Neocomplete and CtrlP
The .erb linter isn't working properly and giving false positives because of a problem with Ruby itself rather than with Ale.  Reported it to the Ale dev here: https://github.com/w0rp/ale/issues/580.

Added Neocomplete but haven't had any time to play with the actual features and been spending more time trying to get used to CtrlP.  Some rebinding will be needed to make it easier to use.